+++ docs/changelog.rst
@@ -5,6 +5,7 @@ v1.3.0
 ------
 Features
   * Close `#12`_: Adding ids to comments so they can be linked to
+  * Close `#28`_: Using django-admin-sortable to allow for re-ordering of topics in the admin interface
   * Close `#29`_: Adding the option to use elasticsearch for searching
 
 v1.2.0
@@ -37,4 +38,5 @@ Bug Fixes:
 .. _#23: https://github.com/smalls12/django_simple_forums/issues/23
 .. _#24: https://github.com/smalls12/django_simple_forums/issues/24
 .. _#25: https://github.com/smalls12/django_simple_forums/issues/25
+.. _#28: https://github.com/smalls12/django_simple_forums/issues/28
 .. _#29: https://github.com/smalls12/django_simple_forums/issues/29